Redesign grammar for compound types
The previous grammar added the meta/data/value to the last item of an And/Or production.  But
we actually want the meta/children/value to be bound to the prototype itself.  The only way to make
this work is to restrict the items of the and/or productions to be simpler.  So we now require:
  - and: only qnames
  - or: only qname + value strings
  - maybe: only qnames